fix : change count type from int to long
- use long to avoid overflow.
e.g
N≦200,000
100 yen -> 100,000.. a
400 yen -> 100,000...b
a * b = 10,000,000,000
10,000,000,000 > int max (4,294,967,295)

Issue: 018 - Convenience Store 1

+++ exercises/chapter3-3/ConvenienceStore1.java
@@ -5,10 +5,10 @@ class Main {
       try (Scanner scanner = new Scanner(System.in)) {
          scanner.nextInt();
 
-         int countOf100 = 0;
-         int countOf200 = 0;
-         int countOf300 = 0;
-         int countOf400 = 0;
+         long countOf100 = 0;
+         long countOf200 = 0;
+         long countOf300 = 0;
+         long countOf400 = 0;
          while (scanner.hasNext()) {
             switch (scanner.nextInt()) {
                case 100:
